refactor: change 'OfferSortingOption' enum (simplified usage)
