+++ project/src/types/types.ts
@@ -1,5 +1,3 @@
-import { OfferSortingVariant } from 'src/consts/consts';
-
 export type Offer = {
   bedrooms: number;
   city: City;
@@ -70,5 +68,3 @@ export type Review = {
 }
 
 export type Reviews = Review[]
-
-export type OfferSortingOption = keyof typeof OfferSortingVariant;
