+++ project/src/utils/utils.ts
@@ -1,5 +1,5 @@
-import { OfferSortingVariant, RATING_TO_PERCENT_STEP } from 'src/consts/consts';
-import { Offers, OfferSortingOption } from 'src/types/types';
+import { OfferSortingOption, RATING_TO_PERCENT_STEP } from 'src/consts/consts';
+import { Offers } from 'src/types/types';
 
 export function getPercentFromRating(rating: number): string {
   const roundedPercent = Math.round(rating) * RATING_TO_PERCENT_STEP;
@@ -78,17 +78,19 @@ export function parseInteger(numberAsString = ''): number | typeof NaN {
 export function sortOffers(offers: Offers, sortingType: OfferSortingOption): Offers {
   const newOffers = [...offers];
   switch (sortingType) {
-    case OfferSortingVariant.popular.id: // FIXME: change to constant
+    case OfferSortingOption.Popular:
       return offers;
-    case OfferSortingVariant.priceLowToHigh.id:
+    case OfferSortingOption.PriceLowToHigh:
       newOffers.sort((offer1, offer2) => offer1.price < offer2.price ? -1 : 1);
       break;
-    case OfferSortingVariant.priceHighToLow.id:
+    case OfferSortingOption.PriceHighToLow:
       newOffers.sort((offer1, offer2) => offer1.price > offer2.price ? -1 : 1);
       break;
-    case OfferSortingVariant.topRatedFirst.id:
+    case OfferSortingOption.TopRatedFirst:
       newOffers.sort((offer1, offer2) => offer1.rating > offer2.rating ? -1 : 1);
       break;
+    default:
+      throwErrorAtImpossibleCase(sortingType);
   }
   return newOffers;
 }
